Move handling of a stream to StreamHandler
